% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_polygon.R
\name{render_polygon}
\alias{render_polygon}
\title{Render polygon}
\usage{
render_polygon(n, polygon, type = "segments")
}
\arguments{
\item{n}{The number of elements used for the rendering}

\item{polygon}{A data frame with x and y coordinates for the vertices}

\item{type}{The type of rendering to be applied (currently only one option)}
}
\description{
Creates renderings that can be added as layers to polygon plots
}
\examples{
render_polygon()
}
\keyword{theme}
